+++ src/components/Loading/SkeletonFestivalSentenceItem.tsx
@@ -1,0 +1,46 @@
+// import { styled } from 'styled-components';
+
+// const SkeletonItem = styled.li`
+//     display: flex;
+//     align-items: center;
+//     position: relative;
+//     overflow: hidden;
+//     height: 100%;
+//   `;
+
+// const SkeletonImg = styled.div`
+//     width: 50px;
+//     height: 50px;
+//     border-radius: 50%;
+//     background: #f2f2f2;
+//     position: relative;
+//     overflow: hidden;
+//     margin-left: 16px;
+//     margin-right: 10px;
+// `;
+
+// const SkeletonInfo = styled.div`
+//   width: 100%;
+// `;
+
+// const SkeletonName = styled.div`
+//     width: 50%;
+//     height: 15px;
+//     border-radius: 10px;
+//     background: #f2f2f2;
+//     position: relative;
+//     overflow: hidden;
+//     margin-top: 2.5px;
+// `;
+
+// export default function SkeletonFestivalSentence() {
+//   return (
+//     <SkeletonItem>
+//       <SkeletonImg />
+//       <SkeletonInfo>
+//         <SkeletonName />
+//         <SkeletonName />
+//       </SkeletonInfo>
+//     </SkeletonItem>
+//   );
+// }
