--- src/components/Loding/Skeleton.ts
@@ -1,15 +1,0 @@
-import { styled } from 'styled-components';
-
-const SkeletonLineUp = {
-  SkeletonImg: styled.div`
-        position: relative;
-        overflow: hidden;
-        background-color: #f8f8fa;
-        width: 220px;
-        height: 21rem;
-        border-radius: 0.9rem;
-        box-shadow: 0px 2px 20px 0px rgba(0, 71, 201, 0.15);
-    `,
-};
-
-export default SkeletonLineUp;
